+++ notebooks/aten.py
@@ -15,8 +15,8 @@
 # %%
 import torch
 
+from aioway._torch import *
 from aioway._utils import *
-from aioway._tensors import *
 from aioway.modes import *
 
 # %%
