+++ notebooks/print_torch_ops.py
@@ -16,7 +16,7 @@
 import torch
 
 # %%
-from aioway._tensors import torch_fake_mode
+from aioway._torch import torch_fake_mode
 from aioway.modes import PrintTorchDisp, PrintTorchFunc
 
 # %%
