+++ src/aioway/_torch/__init__.py
@@ -6,3 +6,4 @@ from .fake import *
 from .guards import *
 from .hooks import *
 from .renders import *
+from .schemas import *
