+++ src/aioway/_torch/fake.py
