+++ src/aioway/_torch/hooks.py
