+++ src/aioway/_torch/schemas/__init__.py
@@ -1,13 +1,8 @@
 # Copyright (c) AIoWay Authors - All Rights Reserved
 
 from .attrs import *
-from .data import *
 from .devices import *
-from .dims import *
 from .dtypes import *
-from .gym import *
 from .layouts import *
-from .media import *
+from .schemas import *
 from .shapes import *
-from .tdicts import *
-from .tensors import *
