+++ src/aioway/_torch/schemas/_bases.py
