+++ src/aioway/_torch/schemas/attrs.py
@@ -10,16 +10,12 @@ from collections import abc as cabc
 
 import torch
 
-from aioway._tensors import torch_fake_mode
-
-from ..spaces import space_dcls
 from .devices import Device, DeviceLike
 from .dtypes import DType, DTypeLike
 from .layouts import Layout, LayoutLike
 from .shapes import Shape, ShapeLike
-from .tensors import TensorSpace
 
-__all__ = ["Attr", "AttrLike", "AttrSpace"]
+__all__ = ["Attr", "AttrLike"]
 
 
 LOGGER = logging.getLogger(__name__)
@@ -73,15 +69,6 @@ class Attr:
 
         if not isinstance(self.shape, Shape):
             raise TypeError(type(self.shape))
-
-    def __space__(self) -> AttrSpace:
-        return AttrSpace(
-            shape=self.shape,
-            dtype=self.dtype,
-            device=self.device,
-            layout=self.layout,
-            requires_grad=self.requires_grad,
-        )
 
     def __bool__(self):
         return True
@@ -141,6 +128,8 @@ class Attr:
         """
         Generate a random tensor. This should be used under fake mode.
         """
+
+        from ..fake import torch_fake_mode
 
         with torch_fake_mode():
             return torch.zeros(
@@ -247,78 +236,3 @@ class AttrLikeDict(typing.TypedDict):
     device: typing.NotRequired[DeviceLike]
     requires_grad: typing.NotRequired[bool]
     layout: typing.NotRequired[LayoutLike]
-
-
-@space_dcls
-class AttrSpace(TensorSpace):
-    """
-    Tag that specify `Attr`s that should be respected.
-    """
-
-    _: dcls.KW_ONLY
-
-    shape: Shape | None = None
-    "The given shape, if given, that must match."
-
-    dtype: DType | None = None
-    "The given dtype, if given, that must match."
-
-    device: Device | None = None
-    "The given device, if given, that must match."
-
-    layout: Layout | None = None
-    "The given layout, if given, that must match."
-
-    requires_grad: bool | None = None
-    "The given requires_grad, if given, that must match."
-
-    def _check_attr(self, attr: Attr) -> None:
-        def check_if_not_none[T](left: T | None, right: T):
-            if left is None:
-                return
-
-            if left != right:
-                raise ValueError
-
-        check_if_not_none(self.shape, attr.shape)
-        check_if_not_none(self.dtype, attr.dtype)
-        check_if_not_none(self.device, attr.device)
-        check_if_not_none(self.layout, attr.layout)
-        check_if_not_none(self.requires_grad, attr.requires_grad)
-
-    def _sample_n(self, batch_size: int) -> torch.Tensor:
-        attr = self.to_attr()
-        attr = dcls.replace(attr, shape=Shape.parse(batch_size, *attr.shape))
-        return attr.to_fake_tensor()
-
-    def _check_data(self, tensor: torch.Tensor) -> None:
-        pass
-
-    def to_attr(self) -> Attr:
-        "Convert to `Attr`. If not enough info, will raise `TypeError`."
-
-        schema: typing.Any = {}
-
-        def add_if_not_none(key, val):
-            if val is None:
-                return
-
-            schema[key] = val
-
-        add_if_not_none("shape", self.shape)
-        add_if_not_none("dtype", self.dtype)
-        add_if_not_none("device", self.device)
-        add_if_not_none("layout", self.layout)
-        add_if_not_none("requires_grad", self.requires_grad)
-
-        return Attr.parse(schema)
-
-    @classmethod
-    def from_attr(cls, attr: Attr, /) -> typing.Self:
-        return cls(
-            shape=attr.shape,
-            device=attr.device,
-            dtype=attr.dtype,
-            layout=attr.layout,
-            requires_grad=attr.requires_grad,
-        )
