+++ src/aioway/_torch/schemas/devices.py
