+++ src/aioway/io/sinks.py
@@ -7,7 +7,6 @@ import dataclasses as dcls
 import typing
 
 from aioway.relalg import Exec
-from aioway.spaces import AnyDataSpace, DataSpace
 
 __all__ = ["Sink", "sink_dcls"]
 
@@ -34,9 +33,6 @@ class Sink[T = typing.Any](abc.ABC):
                 raise TypeError(f"The batch has {type(batch)=}, expected {self.TYPE}.")
 
             self.write(batch)
-
-    def __space__(self) -> DataSpace:
-        return AnyDataSpace()
 
     @abc.abstractmethod
     def write(self, batch: T, /) -> None:
