+++ src/aioway/modes/aten/fn.py
@@ -6,7 +6,7 @@ import dataclasses as dcls
 import logging
 import typing
 
-from aioway._tensors import is_aten_op
+from aioway._torch import is_aten_op
 
 from .aten import Aten, find_aten
