+++ src/aioway/modes/hists.py
@@ -10,10 +10,9 @@ from collections import abc as cabc
 
 import torch
 
-from aioway._tensors import is_leaf_has_grad, replace_tensors_with_attr
 from aioway._thunks import Thunk
+from aioway._torch import Attr, is_leaf_has_grad, replace_tensors_with_attr
 from aioway._utils import TensorInput, find_nested_tensors, topo_sort
-from aioway.spaces import Attr
 
 from .modes import ModeThunk
