+++ src/aioway/relalg/execs.py
@@ -9,7 +9,7 @@ import typing
 from collections import abc as cabc
 
 from aioway._api import public_api
-from aioway._tensors import torch_fake_mode
+from aioway._torch import torch_fake_mode
 from aioway._utils import decomp_dcls_members
 
 from .nodes import GraphNode
