+++ src/aioway/spaces/__init__.py
@@ -1,6 +1,10 @@
 # Copyright (c) AIoWay Authors - All Rights Reserved
 
+from .attrs import *
 from .casts import *
-from .data import *
-from .modules import *
+from .dims import *
+from .gym import *
+from .media import *
 from .spaces import *
+from .tdicts import *
+from .tensors import *
