+++ src/aioway/spaces/gym.py
@@ -4,12 +4,10 @@ import typing
 
 import torch
 
+from aioway._torch import Attr, DType, Shape
 from aioway.errors import re_raise_func
 
-from ..spaces import space_dcls
-from .attrs import Attr
-from .dtypes import DType
-from .shapes import Shape
+from .spaces import space_dcls
 from .tensors import TensorSpace
 
 __all__ = ["DiscreteSpace", "BoxSpace", "MultiDiscreteSpace", "MultiBinarySpace"]
