+++ src/aioway/spaces/tensors.py
@@ -8,10 +8,10 @@ import typing
 import torch
 
 from aioway._api import public_api
-from aioway._tensors import is_fake_tensor
+from aioway._torch import is_fake_tensor
+from aioway.spaces import Space
 
-from ..spaces import space_dcls
-from .data import DataSpace
+from .spaces import Space, space_dcls
 
 if typing.TYPE_CHECKING:
     from .attrs import Attr
@@ -21,7 +21,7 @@ __all__ = ["TensorSpace"]
 
 @public_api
 @space_dcls
-class TensorSpace(DataSpace[torch.Tensor], abc.ABC):
+class TensorSpace(Space[torch.Tensor], abc.ABC):
     "A `Space` that enforces constraints on a `torch.Tensor`."
 
     @typing.override
